+++ src/modules/options.h
@@ -43,6 +43,7 @@
 #include "modules/swap/option.h"
 #include "modules/terminal/option.h"
 #include "modules/terminalfont/option.h"
+#include "modules/terminalsize/option.h"
 #include "modules/theme/option.h"
 #include "modules/title/option.h"
 #include "modules/uptime/option.h"